Move folder security confirmation and usage metrics calculation to cli_utils
