+++ openhands/core/cli_utils.py
@@ -1,36 +1,116 @@
-VERIFIED_PROVIDERS = ['openai', 'azure', 'anthropic', 'deepseek']
+from pathlib import Path
 
-VERIFIED_OPENAI_MODELS = [
-    'gpt-4o',
-    'gpt-4o-mini',
-    'gpt-4-turbo',
-    'gpt-4',
-    'gpt-4-32k',
-    'o1-mini',
-    'o1',
-    'o3-mini',
-    'o3-mini-2025-01-31',
-]
+import toml
+from prompt_toolkit import print_formatted_text
+from prompt_toolkit.shortcuts import clear, print_container
+from prompt_toolkit.widgets import Frame, TextArea
 
-VERIFIED_ANTHROPIC_MODELS = [
-    'claude-2',
-    'claude-2.1',
-    'claude-3-5-sonnet-20240620',
-    'claude-3-5-sonnet-20241022',
-    'claude-3-5-haiku-20241022',
-    'claude-3-haiku-20240307',
-    'claude-3-opus-20240229',
-    'claude-3-sonnet-20240229',
-    'claude-3-7-sonnet-20250219',
-]
+from openhands.core.cli_input import (
+    cli_confirm,
+)
+from openhands.core.cli_output import (
+    COLOR_GREY,
+    UsageMetrics,
+)
+from openhands.events.event import Event
+from openhands.llm.metrics import Metrics
 
 
-def is_number(char):
-    return char.isdigit()
+def check_folder_security_agreement(current_dir):
+    is_trusted = manage_openhands_file(current_dir)
+
+    if not is_trusted:
+        security_frame = Frame(
+            TextArea(
+                text=(
+                    f' Do you trust the files in this folder?\n\n'
+                    f'   {current_dir}\n\n'
+                    ' OpenHands may read and execute files in this folder with your permission.'
+                ),
+                style=COLOR_GREY,
+                read_only=True,
+                wrap_lines=True,
+            ),
+            style=f'fg:{COLOR_GREY}',
+        )
+
+        clear()
+        print_container(security_frame)
+        print_formatted_text('')
+
+        confirm = (
+            cli_confirm('Do you wish to continue?', ['Yes, proceed', 'No, exit']) == 0
+        )
+
+        if confirm:
+            manage_openhands_file(current_dir, add_to_trusted=True)
+
+        return confirm
+
+    return True
 
 
-def split_is_actually_version(split):
-    return len(split) > 1 and split[1] and split[1][0] and is_number(split[1][0])
+def manage_openhands_file(folder_path=None, add_to_trusted=False):
+    openhands_file = Path.home() / '.openhands.toml'
+    default_content: dict = {'trusted_dirs': []}
+
+    if not openhands_file.exists():
+        with open(openhands_file, 'w') as f:
+            toml.dump(default_content, f)
+
+    if folder_path:
+        with open(openhands_file, 'r') as f:
+            try:
+                config = toml.load(f)
+            except Exception:
+                config = default_content
+
+        if 'trusted_dirs' not in config:
+            config['trusted_dirs'] = []
+
+        if folder_path in config['trusted_dirs']:
+            return True
+
+        if add_to_trusted:
+            config['trusted_dirs'].append(folder_path)
+            with open(openhands_file, 'w') as f:
+                toml.dump(config, f)
+
+        return False
+
+    return False
+
+
+def update_usage_metrics(event: Event, usage_metrics: UsageMetrics):
+    if not hasattr(event, 'llm_metrics'):
+        return
+
+    llm_metrics: Metrics | None = getattr(event, 'llm_metrics', None)
+    if not llm_metrics:
+        return
+
+    cost = getattr(llm_metrics, 'accumulated_cost', 0)
+    usage_metrics.total_cost += cost if isinstance(cost, float) else 0
+
+    token_usage = getattr(llm_metrics, 'accumulated_token_usage', None)
+    if not token_usage:
+        return
+
+    prompt_tokens = getattr(token_usage, 'prompt_tokens', 0)
+    completion_tokens = getattr(token_usage, 'completion_tokens', 0)
+    cache_read = getattr(token_usage, 'cache_read_tokens', 0)
+    cache_write = getattr(token_usage, 'cache_write_tokens', 0)
+
+    usage_metrics.total_input_tokens += (
+        prompt_tokens if isinstance(prompt_tokens, int) else 0
+    )
+    usage_metrics.total_output_tokens += (
+        completion_tokens if isinstance(completion_tokens, int) else 0
+    )
+    usage_metrics.total_cache_read += cache_read if isinstance(cache_read, int) else 0
+    usage_metrics.total_cache_write += (
+        cache_write if isinstance(cache_write, int) else 0
+    )
 
 
 def extract_model_and_provider(model):
@@ -79,3 +159,38 @@ def organize_models_and_providers(models):
         result[key]['models'].append(model_id)
 
     return result
+
+
+VERIFIED_PROVIDERS = ['openai', 'azure', 'anthropic', 'deepseek']
+
+VERIFIED_OPENAI_MODELS = [
+    'gpt-4o',
+    'gpt-4o-mini',
+    'gpt-4-turbo',
+    'gpt-4',
+    'gpt-4-32k',
+    'o1-mini',
+    'o1',
+    'o3-mini',
+    'o3-mini-2025-01-31',
+]
+
+VERIFIED_ANTHROPIC_MODELS = [
+    'claude-2',
+    'claude-2.1',
+    'claude-3-5-sonnet-20240620',
+    'claude-3-5-sonnet-20241022',
+    'claude-3-5-haiku-20241022',
+    'claude-3-haiku-20240307',
+    'claude-3-opus-20240229',
+    'claude-3-sonnet-20240229',
+    'claude-3-7-sonnet-20250219',
+]
+
+
+def is_number(char):
+    return char.isdigit()
+
+
+def split_is_actually_version(split):
+    return len(split) > 1 and split[1] and split[1][0] and is_number(split[1][0])
